+++ new_code/wrapper.cu
@@ -28,13 +28,19 @@ using std::cout;
 // ----- BEGIN MAIN PROGRAM -----
 int main(int argc, char *argv[]) {
   const char *fileIn = "check2.lvm"; // file name to load
-  const char *pushFile =
-      "processed_push.csv"; // file name to which we will write the "push" data
-      /*
-  const char *releaseFile =
-      "processed_release.csv"; // file name to which we will
-                               // write the "release" data
-                               */
+
+  // filenames for checking data
+  //const char *filtAcc1File = "filtered_acc1.csv";
+  //const char *filtAcc2File = "filtered_acc2.csv";
+  //const char *pushPullIndicesFile = "push_pull_indices.csv";
+
+  // file name to which we will write the "push" data
+  const char *pushFile = "processed_push.csv";
+  /*
+const char *releaseFile =
+  "processed_release.csv"; // file name to which we will
+                           // write the "release" data
+                           */
   float **dataMatrix;
 
   // number of fields (i.e. data columns) in the .lvm file
@@ -71,14 +77,25 @@ int main(int argc, char *argv[]) {
   cudaMallocManaged((void **)&filteredAcc1, sizeof(float) * numSamples);
   cudaMallocManaged((void **)&filteredAcc2, sizeof(float) * numSamples);
 
+  // check that the files were read correctly
+  // cout << "Taps: " << dataMatrix[0][2999999] << "\n";
+  // cout << "Acc1: " << dataMatrix[1][2999999] << "\n";
+  // cout << "Acc2: " << dataMatrix[2][2999999] << "\n";
+
   filtfilt(dataMatrix[1], filteredAcc1, numSamples, sampleRate, filter[0],
            filter[1]); // filter acc1 data
   filtfilt(dataMatrix[2], filteredAcc2, numSamples, sampleRate, filter[0],
            filter[1]); // filter acc2 data
 
+  // check the filtered signals
+  //cout << "Acc1: " << filteredAcc1[99] << "\n";
+  //cout << "Acc2: " << filteredAcc2[99] << "\n";
+  //writeCSV(filtAcc1File, filteredAcc1, numSamples, 1); // write out push data
+  //writeCSV(filtAcc2File, filteredAcc2, numSamples, 1); // write out push data
+
   /// sort the push and pull tap indices
   size_t *pushPullIndices;
-  int nLead, nTrail;
+  size_t nLead, nTrail;
 
   // returns indices of "push" (rising edges) and "release" (falling edges) in
   // array
@@ -86,6 +103,9 @@ int main(int argc, char *argv[]) {
   // release2, ... releaseN]
   pushPullIndices =
       sort(dataMatrix[0], sampleRate, numSamples, 100, &nLead, &nTrail);
+
+  // check pushPullIndices
+  //writeCSVInt(pushPullIndicesFile, pushPullIndices, (nLead+nTrail), 1);
 
   // compute wave speed
   bool whichFirst = 0;
@@ -163,20 +183,20 @@ int main(int argc, char *argv[]) {
 
   computeWaveSpeed(filteredAcc1, filteredAcc2, ind1, ind2, push, nPush, window,
                    sampleRate, travelDist);
-/*
-  // RELEASE
-  for (i = 0; i < nRelease; i++) {
-    if (!whichFirst) {
-      ind1[i] = pushPullIndices[nPush + i];
-      ind2[i] = pushPullIndices[i + 1];
-    } else {
-      ind1[i] = pushPullIndices[nPush + i];
-      ind2[i] = pushPullIndices[i];
-    } // end conditional whichFirst
-  }   // end for each release
-  computeWaveSpeed(filteredAcc1, filteredAcc2, ind1, ind2, release, nRelease,
-                   window, sampleRate, travelDist);
-                   */
+  /*
+    // RELEASE
+    for (i = 0; i < nRelease; i++) {
+      if (!whichFirst) {
+        ind1[i] = pushPullIndices[nPush + i];
+        ind2[i] = pushPullIndices[i + 1];
+      } else {
+        ind1[i] = pushPullIndices[nPush + i];
+        ind2[i] = pushPullIndices[i];
+      } // end conditional whichFirst
+    }   // end for each release
+    computeWaveSpeed(filteredAcc1, filteredAcc2, ind1, ind2, release, nRelease,
+                     window, sampleRate, travelDist);
+                     */
 
   // ----- timing -----
   // stop the event and get the elapsed time
@@ -187,9 +207,9 @@ int main(int argc, char *argv[]) {
   std::cout << "Elapsed Time: " << elapsedTime << "\n";
   // ----- timing -----
 
-  //cout << "push[0]: " << push[0] << "\n";
+  // cout << "push[0]: " << push[0] << "\n";
 
-  writeCSV(pushFile, push, nPush, 1);          // write out push data
+  writeCSV(pushFile, push, nPush, 1); // write out push data
   // writeCSV(releaseFile, release, nRelease, 1); // write out release data
 
   // free all allocated memory
